Move out vg ids
